Create functions_310.py
