+++ tests/functions.py
@@ -1,5 +1,10 @@
 # ruff: noqa
 # ruff must not change the AST of the test functions, even if they are semantically equivalent.
+import sys
+if sys.version_info >= (3, 10):
+    from .functions_310 import functions_310, unsupported_functions_310
+else:
+    functions_310 = []
 
 
 def signum(x):
@@ -9,18 +14,6 @@ def signum(x):
     elif x < 0:
         s = -1
     return s
-
-def match_signum(x):
-    s = 0
-    match x:
-        case 0:
-            s = 1
-        case 2:
-            s = -1
-        case 3:
-            s = 0
-    return s
-
 
 def signum_no_default(x):
     if x > 0:
@@ -41,16 +34,6 @@ def nested_partial_return_with_assignments(x):
     else:
         return -5 - x
     return s * x
-
-def match_nested_partial_return_with_assignments(x):
-    match x:
-        case 0:
-            return -5 - x
-        case 1:
-            return 1 * x
-        case 2:
-            return 2 + x
-    return -1 * x
 
 
 def early_return(x):
@@ -104,15 +87,6 @@ def compare_expr(x):
         s = 2
     return s
 
-def match_compare_expr(x):
-    match x:
-        case 0:
-            return 2
-        case 1:
-            return 1
-        case 10:
-            return 2
-    return 1
 
 def bool_op(x):
     if (0 < x) and (x < 10):
@@ -224,18 +198,6 @@ def nested_if_else(x):
         s = 0
     return s
 
-def nested_match(x):
-    match x:
-        case 0:
-            match x:
-                case 0:
-                    return 1
-                case 1:
-                    return 2
-            return 3
-        case 1:
-            return 4
-    return 5
 
 def nested_if_else_expr(x):
     if x > 0:
@@ -259,17 +221,6 @@ def assignments_inside_branch(x):
     else:
         s = 0
     return s
-
-
-def match_assignments_inside_branch(x):
-    match x:
-        case 0:
-            return 0
-        case 1:
-            return 2 * x
-        case 2:
-            return 3 * x
-    return x
 
 
 def override_default(x):
@@ -305,27 +256,7 @@ def tuple_assignments(x):
 def list_assignments(x):
     [a, b] = 1, x
     return x + a + b
-
-def match_case(x):
-    s = 0
-    match x:
-        case 0:
-            s = 1
-        case 2:
-            s = -1
-        case _:
-            s = 0
-    return s
-
-def match_with_or(x):
-    match x:
-        case 0 | 1:
-            return 0
-        case 2:
-            return 2 * x
-        case 3:
-            return 3 * x
-    return x
+        
 
 functions = [
     signum,
@@ -354,13 +285,7 @@ functions = [
     multiple_if,
     return_unconditional_constant,
     return_conditional_constant,
-    nested_match,
-    match_assignments_inside_branch,
-    match_signum,
-    match_nested_partial_return_with_assignments,
-    match_compare_expr,
-    match_case,
-    match_with_or,
+    *functions_310,
 ]
 
 xfail_functions = [
@@ -378,4 +303,5 @@ unsupported_functions = [
     (return_end, "return needs a value"),
     (no_return, "Not all branches return"),
     (return_nothing, "return needs a value"),
+    *unsupported_functions_310,
 ]
